test: Fix the selenium test
Fix the test of selenium and add a new file for them
